move navigation logic into a hook and fix formatting
