+++ SUMMARY.md
@@ -231,7 +231,7 @@
 - [NOVA SMS](libs/nova-sms/README.md)
 - [Plausible](libs/plausible/README.md)
 - [Portals](libs/portals/README.md)
-    - [Portals Admin Core](libs/portals/admin/core/README.md)
+  - [Portals Admin Core](libs/portals/admin/core/README.md)
   - [Core](libs/portals/core/README.md)
 - [React Libraries](libs/react/README.md)
   - [Feature Flags](libs/react/feature-flags/README.md)
